Don't use the default Cxx rulekey for a genrule target if one is specified
Summary: When implementing the aggregation for swift module map headers, we used the defaultCxxPlatform which I beleived to be for iphonesimulator-x86_64, but it is actually a buck-wide value that can be configured that way and, thus, changed. This diff changes that to use the defaultPlatform specified on the target for the build flavor for the genrule files.

shipit-source-id: 88b6614c12ff739371763760a648f33677a32e25