Use custom GID generator for new project format
Summary:
This removes using the thread-unsafe class `GidGenerator` altogether in favor of a simpler method.

`GidGenerator` used two pieces of a PBX object to create a "unique" hash -- the PBX object `isa` string name as a hash code and a `stable hash` int which was usually 0 except in cases of PBXReference types, which had a name string hash. To guarantee uniqueness, we also used a ` counter` variable which was incremented.

Given that we want to parallelize a bunch of stuff, it doesn't really matter to us if objects across different runs of buck project are the same; it's a lot slower to check if a GID exists in a hashmap than it is to just generate one and continue.

To solve this, I'm removing the use of GidGenerator in the new format GID generation and simplifying a few things. First, we will skip the  `stable hash` part of the GID generator (since it's almost always 0 anyway) and just making it `0`. Next, the new formula for GID generation will be:

`PBXObject::isa().hashCode() + 0 + {PBXObject::isa() counter reference}` where the counter is incremented for each individual type.

This means that if we have two PBXFileReference types, the counter for the first file will be `0`, the next `1`, and so on and so forth.

This has a considerable performance improvement; generating a large workspace of 10000+ targets goes from ~14 mins (not including codegen part of buck project) to ~3.5mins (not including codegen).

Reviewed By: williamtwilson

shipit-source-id: 9460428036f4b2a80b8514508cef2d5b3d9e7505