Move some file writes out of nativeTargetGenerator for thready safety
Summary:
I believe that writing to the filesystem is not thread safe, and seeming sporadic crashes writing to disk after a `buck clean`. From what I can tell, two items are trying to write to the same directory but when the second one tries to, it's crashing because the directory is not empty (apparently it's trying to delete the existing stuff before hand).

To resolve this, I'm moving the private symlink writes out of the futures part and doing it during the processing of the targets. This means we pass around some metadata back out, but then are able to write them later on. This doesn't seem to have much of a performance hit from my testing.

shipit-source-id: 17fb659ac05bf3571d6a8a7173118100f05214ce