Migrate new PBXObject* call sites to Factory
Summary: This updates the callsites of projectv2 to use the new project factory that will create and assign GIDs when a PBX* object is created. This means that by the time we get to serialization, all objects will have a GID assigned and we can treat each one as immutable (even tho they technically are mutable still because of backwards compat with projectv1).

Reviewed By: williamtwilson

shipit-source-id: c1d7272cc4932dd60987ddf212ce8f4ab7cefbd2