dogfood project.buck_out_include_target_config_hash=true
Summary:
We need to enable this option in repositories, but before that it would be helpful to test it in our repository.

Note: this diff only affects Buck repository, but not other repositories.

Note: this diff only enables this option for Buck project, but not for tests running by Buck, D19374631 should enable running tests with this option on (these changes are independent).

Reviewed By: cjhopman, bobyangyf, gabrielrussoc

shipit-source-id: b87968d8857125bd6196425e6f1bb30503814ef6