Change how Args rulekeys are computed
Summary:
Annotate the appropriate fields rather than a method in some parent
class. Rulekeys should be computed based on the referenced data, not a custom
method, the method support is intended only to be used by immutables to
annotate their data methods (which explains the otherwise useless
BuckStyleImmutable annotation on CommandLineArgs).

Reviewed By: philipjameson

shipit-source-id: a5c008ff4083b26f8d8aab761a7c4f5c6fd27616