Support coercing output label in BuildTargetMacroTypeCoercer
Summary:
This is in preparation for getting `OutputLabel` out of `BuildTargetMacro`, which will be done by replacing `getBuildTarget()` with `getBuildTargetWithOutputs()` in `BuildTargetMacro`. We want this so it's easier for the subclasses of `BuildTargetMacro` to support named outputs

The change in `TargetNodeFactory` is to make sure that the coerced object from the macro coercer still gets added to target graph. Don't need to keep the output label info in the graph because that's resolved separately

Reviewed By: bobyangyf

shipit-source-id: ff9d8602d3beb741ded791b69324d3a4bdbdd48d