BinaryBuildRule should know about output labels
Summary:
`BinaryBuildRule` should know about output labels to know which output to execute if it has multiple named outputs.

To support this, change the signature of
  Tool getExecutableCommand();
to
  Tool getExecutableCommand(Optional<OutputLabel> outputLabel);

Note that when full multiple outputs (i.e. single element list restriction is removed) is supported, we will need to enforce in the method implementation that the `ImmutableSet<OutputPath>` associated with the given `OutputLabel` in the `BinaryBuildRule` actually still refers to only one `OutputPath` (unless we want to allow the user to run all the outputs? Anyway I think that gets complicated quickly and should probably be a separate feature)

Callsite in `ExecutableMacroExpander` will be fixed in follow-up diff to propagate output label.

Checked the different callsites and we probably also care about the callsite in:
* `RunCommand` (it will allow us to do things like `buck run //:foo[label]`)

Other callsites are more language-specific and probably don't need to be changed right now until there's a use case

Reviewed By: bobyangyf

shipit-source-id: f6978416fb0779ed99385e534fea6d4276535152