Allow a non-skylark API to get OutputArtifact
Summary:
Right now if you want to get an OutputArtifact, you have to use a skylark focused API (e.g. that requires a location, and throws an eval exception).
For java rules, this is just boiler plate that needs handled, so add a non-skylark asOutputArtifact to the Artifact API to make java devs' life easier.

Reviewed By: bobyangyf

shipit-source-id: cc03b220c5d81e8df2d8e4e37fcbf83c3affac2a