Use rule name, rather than java class name when bridging to legacy action graph
Summary:
Previously we were setting the name of RuleAnalysisLegacyBuildRuleView to the name of the constructor args' java class. Instead, pull a human readable name out.
This will pull from the DescriptionCache for java rules. For UDR, it will use a slightly different interface, as we need to have the SkylarkDescriptionArgs in order to determine the real rule name

Reviewed By: bobyangyf

shipit-source-id: c0a10651743345499f894cb8072d1131b510daf6