Reduce the amount of items codegen'd at the end of a project run
Summary:
Our changes to how buck project organizes and determines what files to codegen has become too eager. In buck project v1, we only really generated code for projects with targets in them (so basically, an entire package), but if a project was not focused on, we would just discard the targets it needed generated. In projectv2, we are not really doing that -- we are just generating everything which means we are paying a huge cost at the end of buck projectv2.

To fix this, I'm reducing the set of codegen targets to those that are part of the focus set; we don't really need to codegen anything outside of the focus-set because they will never be part of the Xcode project -- and the first call to the buck build script will codegen those files anyway.

To do this, I elevated the work done for these source paths to the WorkspaceAndProjectGenerator level, which contains the focus target matcher, so it can determine which targets need to be built.

The only exception are swift modulemaps; we are just going to naively codegen all of them for now because without which, Swift editing in Xcode tends to be a terrible experience since click-throughs and auto-complete will never work.

Reviewed By: williamtwilson

shipit-source-id: 6f17aeb1d3a7ea4270c5a27f6bdd77c38a73cc42