Fix filtering of workspace target
Summary: We started seeing a duplicate key crash when running projectv2 which D19505569 worked to resolve yesterday with a Set. I've determined this to be a result of the fact that we're not filtering the workspace node from the other target sets. We process the workspace target node first and then shouldn't process again; however our equality check wasn't working so we ended up creating it twice.

Reviewed By: williamtwilson

shipit-source-id: a5f78fa20d65c42f5247d97644658e6db51e6227