Port JarGenrule away from LegacyGenrule
Summary: Same as the other diffs in this stack, this commit ports JarGenrule away from LegacyGenrule and onto the modern Genrule. This genrule is a little weird, though; despite reporting its output as an unsuffixed JAR file, its "actual" output is a file ending in .jar. We can't change this, since $OUT is set to the genrule's "out" argument, and the out argument is always an unsuffixed file. We can work around this by copying the unsuffixed JAR to the suffixed JAR and leaving both files in place. This rule isn't used much and this works fine, so I'm calling it OK.

Reviewed By: irenewchen

shipit-source-id: 17ebafa3860a085e3fcee2d1fd633a95a17abcd7