Port JsBundleGenrule away from LegacyGenrule
Summary:
Like the previous diffs in this stack, this commits ports JsBundleGenrule away from LegacyGenrule and back onto our modern Genrule that we know and love.

This diff is more complicated than the previous ones. This is because JsBundleGenrule is special in that it exports up to four ourputs:

1. A SourceMap for bundled Javascript,
2. A misc folder that contains arbitrary stuff (outside of Buck's control)
3. A dependency file, for keeping track of dependencies between packages
4. A js folder, containing the bundled JavaScript

JsBundleGenrules can inherit these directly from other bundles that it depends on, but it also has rule arguments that allows it to override and mutate the source map, dependency file, or misc folders. Mutating any one of these adds an output to JsBundleGenrule (so it has 1 output if nothing is changed, and 4 if everything is).

The core functionality of JsBundleGenrule is completely extracted into a Buildable implementation. This was mostly straightforward; the only difference between the current implementation and the one in this diff is that the one in this diff inspects its own outputs whenever it needs to mutate an output.

Reviewed By: mykola-semko

shipit-source-id: 9937ea1096cbcd0b0310f6d119650882a54974b0