Update buck to support adb toolchain from PATH
Summary:
Adds a buck config option, `android.adb_search_order`, which is used to determine how to find adb. If set, it's expected to have a comma separated list of places to search, which can include: `<PATH>` to indicate the system PATH environment variable, and `<ANDROID_SDK>` to indicate the `platform-tools/adb` subpath of the configured Android SDK.

If the `android.adb_override` buck config option is set, it takes precedence over `android.adb_search_order`.

Reviewed By: christolliday

shipit-source-id: af139a52b4938f1c96422310e503c47a8e8c3edf