Add new pmd rules about immutables
Summary:
We want people to always use BuckStyleValue variants instead of raw immutables.

We also want overwriting of defaults of the existing immutable styles to specify everything, so that it is only used when special things needs to be changed.
Most of the times, BuckStyleValue variants should be sufficient for the average use.

Reviewed By: stepancheg

shipit-source-id: 585173796fa5b191055e26208f59d667f026c5f8