Fix parsing of lists with empty values
Summary: Previously, when using $(config) we could end up with empty strings after interpolation. This is after all of our trim/isEmpty logic runs, so we might assume a value is present that is actually an empty string and should be ignored. This broke when we used $(config) and got an empty string in a place where a list was expected, and we couldn't run get(0) on an empty set. Return Optional.empty() from getValue() if no non-whitespace values are found, rather than doing .get()

Reviewed By: gnuthor

shipit-source-id: 8db8a1d59ac780a27dd3dee1f23c346a57d3f9f0