Remove bytes_to_str
Summary: In py2, we just wrote a binary blob to stdout/stderr. We can't do that directly in python3, however, we can write to stdout/stderr.buffer with raw bytes. Remove the decode-to-utf8 code, and just write out bytes. Note that this requires an extra 'flush', as stdout/err gets buffered by default and automatic flushing only happens with text output.

Reviewed By: stepancheg

shipit-source-id: 84c697880e1602842f06972c15653919e59d5470