Correct "Default" OSX sharedLibraryVersionedExtensionFormat
Summary:
The DefaultCxxPlatforms.java defines an incorrect sharedLibraryVersionedExtensionFormat string on OSX and has an extra period prefixed to the front of it.  If this code were to be used along with the following rule:

    cxx_library(
        name = "testlib",
        ...
        soname = "libtestlib.$(ext 1.2)",
    }

then a shared library with the name "libtestlib..1.2.dylib" would be produced.  i.e. it would have an extra period in it.

Luckily, the code in DefaultCxxPlatforms.java does not seem to be exercised when building a shared library on OSX.  Instead the code in AppleCxxPlatforms.java, which has always specified a correct format string, is used.

Let's go ahead and remove the extra period from DefaultCxxPlatforms.java to ensure that it does not cause confusion.

Reviewed By: miaoyipu, scottrice

shipit-source-id: b65b93132616f835c209f4c217f6dd0f13b42b8a