Adding console output when waiting for debugger to attach (#2376)
Summary:
We've had a number of developers start working with buck and not grasping that it's suspended and waiting for the debugger to attach.

This added output to inform the developer to attach the debugger.
Pull Request resolved: https://github.com/facebook/buck/pull/2376

Test Plan:
Start build a target in debug mode
NO_BUCKD=1 BUCK_DEBUG_MODE=1 ./bin/buck build <Your target>
You should see the message "Waiting for debugger on port 8888..."
Attach the process in IntelliJ and you are able to debug buck.

Reviewed By: philipjameson, v-jizhang

Pulled By: v-jizhang

shipit-source-id: 186faf241082687bf6d9ae0ad8f0c86d76f3e63e