Create KotlinJavaRuntime library if necessary
Summary: This mimics the behavior of Kotlin plugin and adds the library to every module that uses Kotlin. To enable this, users need to set `intellij.kotlin_java_runtime_library_template_path` to point to a template library file, which will be used to produce the `KotlinJavaRuntime.xml` library file.

Reviewed By: brianduff

shipit-source-id: 98334c081d82aa5569068c7799d146fc07dae72a