Fix dex weight warning for dex groups
Summary:
Since dex groups were introduced, the primary dex weight check has been broken since it will only fail if the primary dex weight for an individual dex group is too high.

D8 will throw an error if the 65k method ref limit has been broken, so this check isn't strictly needed, and it's not as accurate as D8 since it counts bytes instead of method refs, but D8 will not tell you whether it was a primary or secondary dex that overflowed, and it won't give you any information on what the heaviest dexes are. For this reason, this turns the error into a warning.

The following diffs catch the case where D8 detects a method ref overflow and use this information to provide a more useful error without guessing if the ref limit overflowed.

Reviewed By: IanChilds

shipit-source-id: 0385787397dfd4894af149e2e1efe237d6ecdc98