Use STGroup to avoid creating the template parser multiple times.
Summary:
Use STGroup to avoid instantiating templates over and over again. This speeds up generation of large projects.

Although doing this is overall faster, the render() method in ST takes longer when STGroup is used because it increases lock contention on the synchronized ObjectModelAdaptor.getProperty() method (https://github.com/antlr/stringtemplate4/blob/master/src/org/stringtemplate/v4/misc/ObjectModelAdaptor.java#L59) which erodes some of the gains. It's a bit unfortunate, since the parameters we pass into the template engine are immutable, and so the synchronization ST is doing here is unnecessary.

Reviewed By: cute-jumper

shipit-source-id: fcfccb1913a8946d2cc54cc9a7a9f1f7bd9ca27d