AbsPath, RelPath v2
Summary:
This diff is a better design of D19638247.

## Compared to previous version

In this version `AbsPath` and `RelPath` are not classes, but interfaces implemented directly by `BuckUnixPath` subclasses, thus using `AbsPath` and `RelPath` does not has memory or GC overhead (but has a little CPU overhead).

## Why

It's not always easy to answer these questions:

* is `ProjectFilesystem` root always absolute?
* is cell root always absolute?
* this particular rule specifies command line arguments using absolute or relative paths?
* `buckOut` field in `BuckPaths` is absolute or relative or can be both? Is it possible that it is different for old cells and embedded cells?
* are we sure that `relativize` function always return relative path or we need to double check?
* are we sure that `resolve` function always returns absolute path?
* what `ProjectFilesystem.getPath` function returns — absolute path, or relative path, or can return both?
* and so on

I'm a bit obsessed with types, but:

* documentation is OK
* variable names work better than documentation, e. g. `relativePath` for variable name is better than `path`
* assertions work better than names `Preconditions.checkArguments(path.isRelative())`
* but types work even better than assertions (`Path` is better than `String`, but `AbsPath` and `RelPath` is better than `Path` where appropriate)
* (dependent types work even better than plain types, but we don't have them in Java)

We have a lot of broken code when handling paths, for example:
* some rules do not work correctly in cross-cell scenarios
* certain rules do not work correctly when `--isolation-prefix` is specified to a path containing a slash
* a lot of functions inside Buck ruthlessly call `Path.toAbsolutePath` assuming that current directory is what they expect

Introducing `AbsPath` and `RelPath` should make code easier to maintain in Buck use cases where filesystem paths are much more important than in typical program.

## Implementation

* `AbsPath` and `RelPath` are interfaces with `Path getPath()` methods
* `AbsPath.getPath` is guaranteed to return an absolute path object
* `RelPath.getPath` is guaranteed to return an relative path object
* `PathWrapper` is a base interface which shares common implementation and useful in certain scenarios like `ProjectFilesystem.exists`. Can be removed at cost of slight increase of code size.

`AbsPath` and `RelPath` each have two implementations:
* `AbsPathImpl` and `RelPathImpl` which are simple wrapper around `Path`
* `BuckUnixAbsPath` and `BuckUnixRelPath` implement `AbsPath` and `RelPath` directly, thus
* conversion from an absolute `Path` to `AbsPath` for `BuckUnixPath` is just a cast, no allocation
* and accordingly `Path` to `RelPath`  is also just a cast

We use `BuckUnixPath` in production Buck, thus using `AbsPath` and `RelPath` types in interfaces do not create any memory/GC overhead, but create a little CPU overhead because of casting.

## Downsides

These types make code easier to read, but writing code becomes a little harder (e. g. `Files.mkdir` cannot be called directly on `AbsPath` object).

## Following diffs

This diff only introduces types. Following diffs start to use it.

If you feel strongly against this complication, that's fine. I started writing this stack while debugging some rule and trying to make sense of what's going on in there, but we can continue to live with plain `Path` objects and assertions.

Reviewed By: bobyangyf

shipit-source-id: cbb77c202b19689234e68ce753eec3e4b1a65760