Switch Mach-O file hashing to Murmur3
Summary:
We want to use the fastest hash to produce a unique IDs. Those IDs do *not* have to be cryptographically secure, so there's no reason to spend CPU cycles on SHA-1.

On a 745MB binary, it reduces the hashing time from 3.2s to 0.45s (warm file cache; Mac Pro 2013).

shipit-source-id: dfba49caff68fe2beb190f1410453788da0205f7