Change WebSocket interface to formalize default event set
Summary:
Instead of subscribing the listener to ALL events if you don't specify a set, subscribe them to a hardcoded list of default events which matches the current set of all events.

In subsequent diffs I plan to add more events, but I do not want to overload existing clients with traffic they don't care about.

Ideally we'd make a breaking change and only give you events you subscribed to, but I don't have the appetite for that yet.

Reviewed By: mykola-semko

shipit-source-id: 51da73824990f93085de249f8d45311349f62ad4