Detect python3 for unflavored python_binary targets
Summary:
If a specific python platform is not specified, we run 'python --version' to detect the version of python. On systems like newer versions of centos that only have python2 or python3, this is undesirable behavior, and can cause builds to fail. We put python3 at the end so as to not mess up anyone depending on python2 having precedence.

In general python_binary should specify its default platform

Reviewed By: cjhopman

shipit-source-id: c7971fa6d1891945b645b7a632076578ef656b75