BuildStatusEvent: A new WebSocket event type
Summary:
This represents a substantial change to how the WebSocket build event protocol works.

As [originally described](https://fb.workplace.com/groups/buckeng/permalink/1923286791053260/), the protocol is tightly coupled to the Buck event bus. It provides a way to serialize event bus traffic to JSON.

This design has advantages and disadvantages noted in the linked document. One notable disadvantage is that there is no way for newly connected clients to re-synchronize if they connect in the middle of a build. This is problematic for IDE integrations like those for VSCode, IntelliJ, and the "BuckProgress" menu bar app intended for use with Xcode. In practice WebSocket connections drop, or someone may launch the IDE in the middle of a build.

This diff adds a new event `BuildStatus` which is *not* posted to the `BuckEventBus`. It only exists for WebSocket clients. It provides a basic snapshot of build state that is sufficient to render UI equivalent to the summary line of the SuperConsole.

Unlike other events, `BuildStatus` is sent **at most every 500ms** while a build is active. This ensures we don't flood clients with excessively granular updates.

This event is not sent to clients by default; it must be specifically requested by using the `event` parameter when establishing the WebSocket connection. This is intended to avoid excessive WebSocket traffic for clients that do not want this data. See D19857382 for a related change.

If a client connects and subscribes to `BuildStatus` while a build is in progress, they are immediately sent a `BuildStatus` event with a snapshot of the current build state. This ensures that clients instantly re-synchronize even during periods of quiescence (e.g. action graph computation, during which no events may be posted to the event bus for a minute or two).

The tradeoff is that `WebServerBuckEventListener` now has state, and we pay for another Executor thread to manage the timer. These are not tradeoffs I take lightly, but I think on balance this is worth it.

Reviewed By: swgillespie

shipit-source-id: fd974545b537f8e6c0f342bb9910f077a0964129