Add Reason to Resumed event
Summary: When Buck fires an event indicating that it has resumed an event, it's not obvious to post-processors of the event stream what Buck intends to do with the rule. This commit adds a 'reason' field to Resumed that is serialized to the machine log and indicates what Buck was intending to do when it resumed a rule. This commit only adds an 'UNKNOWN' reason and doesn't add it to the 'Resumed' constructor; future diffs will expand on this idea and add reasons for calls to 'Resumed''s constructor.

Reviewed By: rajyengi

shipit-source-id: 07ff43887e2e9e5936cac5f5413c66a51af2984e