Move target node filtering before parsing logic
Summary: When select for some argument fails due to no DEFAULT, we fail the command. However, if it is due to incompatible platform, we can suppress it. We already skip all incompatible rules, but this is after we configure the graph. The main reason why we did this is because of target_compatible_with which itself requires select. However, now we have removed target_compatible_with.

Reviewed By: stepancheg

shipit-source-id: 91bca0b75665190a511eb033b9bbeaadfc00b519