SelectorListTypeCoercer is not a coercer
Summary:
Coercer is meant to coerce to the final type of field of constructor arg, and we don't have fields of type `SelectorList`.

Thus `SelectorListTypeCoercer` does not need to be a part of coercion framework. There's not need for it to implement `Coercer` interface and register it as coercer.

Note this diff removes lots of dead code from `SelectorListTypeCoercer`, in particular
* the whole `SelectorListTypeCoercerTest` which was testing that dead code
* `Concatable` implementation for `SelectorListTypeCoercer`

This diff is a part of project of moving coercion to unconfigured target graph.

(Note: this ignores all push blocking failures!)

Reviewed By: KapJI

shipit-source-id: 86b8f43ebb65ac906b5317a59e11c6f1e3e6195e