Fix display of build progress for buck project --show-output
Summary:
I was mystified by why arc focus would hang for minutes at a time while generating the Xcode project. My report is here:
https://fb.workplace.com/groups/967178219989117/permalink/3627312217309024/

Today I decided to dig in. The root issue is that we were doing this sequence of events:

1. Generate Xcode projects.
2. Print paths to Xcode projects to stdout, if `--show-output` was passed.
3. Build any targets that need to be built.

However, writing to stdout silently disables the `RenderingConsole` entirely, thanks to this code:

https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/xplat/build_infra/buck_client/src/com/facebook/buck/event/listener/RenderingConsole.java?commit=5d753e0b7a9289dfa8f57cce8f6397bde9d7f6de&lines=164-174

Oops.

The fix is simple. Change the sequence of events to:

1. Generate Xcode projects.
2. Build any targets that need to be built.
3. Print paths to Xcode projects to stdout, if `--show-output` was passed.

This still probably isn't the ideal fix, as I note in a comment inline. Because we are writing directly to stdout, there may be stale `RenderingConsole` output onscreen and we will not have any chance to write the final state to stderr before we terminate rendering. Oh well.

Reviewed By: cjhopman

shipit-source-id: cb1fd3948dbe8b454fe05e7aedd342046f4c8b3c