Add "headers" modulemap mode
Summary:
This adds a new mode for generating `modulemap` files that will create an explicit `header` declaration for each exported header in the library. For example:

```
module Library {
    header "A.h"
    header "B.h"
    ...
}
```

This should be preferable over the two existing modes that we have:

- Umbrella header: requires umbrella headers to actually exist, so engineers must maintain them manually, or codegen must be used. Also, gets messy when the library contains a header of the same name that isn't an umbrella header, such as a library `FeedViewController` that contains a class `FeedViewController`, but also contains other classes, like `FeedView`, etc.
- Umbrella directory: resolves that issue, by essentially globbing the directory of headers. However, this leads to nondeterministic import order on different machines.

Since Buck knows all of the headers in a given library, it can instead generate a modulemap that explicitly declares each of them. This gives us the best of both worlds: no need to maintain or generate an umbrella header and resolve name conflicts, and deterministic import order across machines.

Reviewed By: milend

shipit-source-id: bb9356ef19a8ba9caf60bd129503cb82d6f629b0