Allow haskell platform to override archive contents config
Summary:
Allow haskell platform to override archive contents, particularlly, use normal
archive instead of thin archive when cxx platform is specified to use thin
archive. As the objects produced by ghc is not fully deterministic as of today,
by overriding this for haskell library, we can enable cache for it.

This change will be a no-op when the config is not overridden. As the
ArchiveContents is part of the rule key, we should be able to safely enable
and disable the config without worrying about pollute the cache.

Reviewed By: mykola-semko

shipit-source-id: 73ae253908c3ed64a863387b026e904fbd5cd31d