Added log statement to make report generated by rulekeydiff.py clearer
Summary:
Changed the script to print "Rulekey divergence detected." when `all_results` is not empty as that shows that there was divergence.

Currently the Sandcastle job that `xplat/scripts/build-perf/rulekeydiff.py` creates, calls `diff_rulekeys.py` as one of the steps and just dumps the output to a text file which then ends up in the task that it creates to report its findings to the user. This task's title and message don't convey any information to the user. Adding this extra print statement will allow us to clearly understand when there is rule key divergence (and when it's not therew) and set the title and message of that task appropriately.

Reviewed By: christolliday

shipit-source-id: ab3e1b397f325ae76017cb27035661ada0865d10