Replace deprecated path to Android resource directory with subdir_glob() (#2277)
Summary:
According to the android_resource [docs](https://buck.build/rule/android_resource.html#res), having a path to a directory containing Android resources is deprecated and should be replaced with `subdir_glob()`.

> **res (defaults to None)**
A dictionary mapping relative resource paths to either the resource files or the build targets that generate them. The **subdir_glob()** function can be used to generate dictionaries based on a directory structure of files checked into the repository. Alternatively, this can be a path to a directory containing Android resources, although this option is **deprecated** and might be removed in the future.

See the examples provided:

```
android_resource(
  name = 'res',
  res = subdir_glob([('res', '**')]),
  package = 'com.example',
)
```
Pull Request resolved: https://github.com/facebook/buck/pull/2277

Reviewed By: mykola-semko, v-jizhang

Pulled By: v-jizhang

shipit-source-id: 02c134f81415278addd868550a6e4758e367d994