Add dex group index to rulekeys
Summary:
Without including the dex group index in rulekeys, it's possible to get cache hits on the final `#r_dot_java_dex` dex group when the predexed inputs to that rule haven't changed but the total number of dex groups has. When this happens, and new dex groups are added, the secondary dexes from the new final group will be overwritten by dexes from the cached `#r_dot_java_dex`.

This only happens with `#r_dot_java_dex` because all other dex group rules include the group index in the rule name, and therefore the rulekey (ex. `#secondary_dexes_100`)

This has manifested as missing `BuildConfig` classes at runtime, since generated `BuildConfig` classes from `android_build_config` rules are added to the end of the list of predexed source library dependencies and end up in the last dex group.

This changes the type of `groupIndex` from an `OptionalInt` to an `Optional<Integer>` because `OptionalInt` is not a supported value type by `RuleKeyBuilder`

Reviewed By: IanChilds

shipit-source-id: 227686e233da5afe1402a4abef8c5c54bada6ac8