SkylarkUserDefinedRuleTest do not rely on ParamInfo equals and hashCode
Summary:
`ParamInfo.equals` compares only by name.

(This by the way this test did not actually compare attributes inside `SkylarkParamInfo`).

This diff is a preparation to remove `equals`, `hashCode` and `compareTo` in `ParamInfo`, which do not make much sense when `ParamInto` is parameterized by type, but compared only by name.

Reviewed By: bobyangyf

shipit-source-id: 38298c65fb259f065d2bd1253f7d7c3ab54b60cb