Create schemes for APP_EXTENSION subtypes (#2348)
Summary:
Buck will not auto-create schemes for app extensions that are subtypes of the `APP_EXTENSION` (`com.apple.product-type.app-extension `) type. Those are, among others, the iMessage apps. (`com.apple.product-type.app-extension.messages`)

To avoid adding all possible extension types to the `ProductTypes`, this PR modifies scheme creation to check if the product type BEGINS with the parent extension type. Open to other suggestions on how to fix this if this is deemed too hacky!
Pull Request resolved: https://github.com/facebook/buck/pull/2348

Reviewed By: swgillespie, irenewchen, v-jizhang

Pulled By: v-jizhang

shipit-source-id: 3e27be782cc6709f5e404a22dd2ff94f20cda858