Remove uploadCompleteFuture
Summary:
This commit ensures that build artifacts get uploaded before finishing builds.

The `uploadCompleteFuture` inside of `CachingBuildRuleBuilder` was created too late to be returned as part of its `BuildResult`. The execution was:

1. Do the build
2. Return a `BuildResult` with the default `uploadCompleteFuture` (`Futures.immediateFuture(Unit.UNIT)`)
3. Asynchronously call `.handleResult`
4. **now** set `uploadCompleteFuture`

On top of this, nobody ever waited on the (unreturned) `uploadCompleteFuture`, so there was never a guarantee that it would actually complete. They would generally be uploaded since a second task would be submitted to the HTTP write executor which was awaited.

The `BuildResult`'s `uploadCompleteFuture` is never used, so we can remove it.

Reviewed By: miaoyipu

shipit-source-id: dd2f6f0aa1457bb7b265095f0563a0e670792019