Reparse extension files from python parser if they include a UDR
Summary:
If the python parser returns a target with `buck.type` pointing to a user defined rule, make sure that we parse that .bzl file with Skylark as well. This is necessary to populate the 'known user defined rule' caches so that we can properly coerce the raw target node to a target node.

This will be utilized in a future diff to the python dsl parser

Reviewed By: bobyangyf

shipit-source-id: a07339a3f61b6fc794decb65c9c229479cf136fc