Disable UDR if RAG is disabled
Summary: If rule analysis is disabled, or we only have a python parser enabled, UDR is useless. Make sure that we only turn it on when those other features are enabled. It allows us to use this setting as a source of truth, rather than having to check other settings elsewhere. (e.g. exposing functions in the python_dsl parser, then not being able to parse those results leads to odd behavior)

Reviewed By: bobyangyf

shipit-source-id: 2a5d0591705f996f50ebe03b27dbf6c66cc745ff