Add `struct` implementation that makes values immutable
Summary: Add a `struct` implementation that makes all of the values passed in immutable. This makes it so that users can pass structs to `provider()` (which requires all values to be frozen). For now this is configurable so that "fix buck" and "fix the repo" can be done in separate steps. This will be a short lived option.

Reviewed By: bobyangyf

shipit-source-id: 8d54e3e1308ed0d4e976d1aff533b904dd378f35