Add a helper method to run tests for both --show-output and --show-outputs
Summary:
Adds a helper method to run tests with both `--show-output` and its `--show-outputs` equivalent. The equivalents are mapped in `SHOW_OUTPUT_TO_SHOW_OUTPUTS`. In follow-up diffs, it will look something like:

  private static final ImmutableMap<String, String[]> SHOW_OUTPUT_TO_SHOW_OUTPUTS =
      ImmutableMap.of(
          "--show-output", new String[] {"--show-outputs"},
          "--show-full-output", new String[] {"--show-outputs", "--output-format", "full"},
          "--show-json-output", new String[] {"--show-outputs", "--output-format", "json"},
          "--show-full-json-output",
              new String[] {"--show-outputs", "--output-format", "full_json"});

The `--show-outputs` equivalents are String arrays for easy concatenation with `ObjectArrays.concat` as well as `System.arraycopy`

Reviewed By: bobyangyf

shipit-source-id: 7d64a93610c2295456e6407030201477218f5419