Write output directly in DirArtifactCache
Summary:
When DirArtifactCache was introduced in D796640, the output was copied to a temp directory and then moved to ensure that we didn't have dodgy artifacts if Buck was interrupted midway through the copy.

However, we now use the metadata to determine whether a valid output is present, so AFAICT we should be able to copy the output to the artifact path directly.

Reviewed By: brianduff

shipit-source-id: 4917b5e047d06843265427dff01a1b54625dce18