introduce extra_filtered_resources arg to android_binary
Summary: Introduce an extra_filtered_resources arg to android_binary to filter out a specific list of resources from the apk. The individual items in the list can contain wildcards. This feature is experimental, and may be removed. Unlike normal resource_filters, this implementation is fast, because it directly removes entries from the apk archive instead of unpacking and repacking it.

Reviewed By: YaoPersonal

shipit-source-id: 12343dea6f9a4cbd4743789aa5c0616310629748